+++ src/dftbp/dftbplus/inputdata.F90
@@ -149,12 +149,9 @@ module dftbp_dftbplus_inputdata
     !> Choice of electronic hamiltonian
     integer :: hamiltonian = hamiltonianTypes%none
 
-  #:if WITH_API
     !> Is this ASI callback interface for H,S,P enabled (Stishenko et al.,
     !! https://doi.org/10.21105/joss.05186)
     logical :: isAsiCallbackEnabled = .false.
-
-  #:endif
 
     !> Random number generator seed
     integer :: iSeed = 0
@@ -235,6 +232,15 @@ module dftbp_dftbplus_inputdata
 
     !> Printout of Mulliken
     logical :: tPrintMulliken = .false.
+
+    !> Print pairwise Mulliken bond populations
+    logical :: writeBondPopul = .false.
+
+    !> Print pairwise non-SCC bond energies
+    logical :: writeBondEnergy = .false.
+
+    !> Print pairwise Mayer bond orders
+    logical :: writeBondOrder = .false.
 
     !> Net atomic charges (i.e. on-site only part of Mulliken charges)
     logical :: tNetAtomCharges = .false.